make_thumbnail/requirements.txt
google-cloud-storage
wand
